Add cast of inputs to `compute_dtype` in `RescalingLayer`. (#272)
Note that this cast exists in the `tf.keras` implementation.

+++ keras_core/layers/preprocessing/rescaling.py
@@ -38,7 +38,7 @@ class Rescaling(Layer):
         dtype = self.compute_dtype
         scale = ops.cast(self.scale, dtype)
         offset = ops.cast(self.offset, dtype)
-        return inputs * scale + offset
+        return ops.cast(inputs, dtype) * scale + offset
 
     def compute_output_shape(self, input_shape):
         return input_shape
